terraform/gcp/modules/agent/files/watcher.py/requirements.txt
google-cloud-compute
google-cloud-scheduler
protobuf